ASoC: msm8952: restore eccho reference dailink
when the ospl feature was been propagated to shm2018 line, this dailink
was brought which is used to send I/V feedback signal to QC DSP. But It was
removed in the below change later. in order to not affecting normal user
case, we should not use cs47l35-slim1 to send I/V feedback or external
EC REF signal.

https://gerrit.mot.com/1203865

Change-Id: I032fb36806c4ad39457e3a1a1aa2746186c62836
Signed-off-by: huangys <huangys@motorola.com>
Reviewed-on: https://gerrit.mot.com/1247754
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Seungyoon Lee <yoon@motorola.com>
Submit-Approved: Jira Key